Ensure ENV tag is committed on PC changing instructions
PC changing instructions result in a TB jump before reaching the
validator commit helper call.  Rather than decoding the instruction
to decide where to place the commit, have the validator helper check
if its partner commit got skipped, and if so run the commit before
the validate.

With this scheme breakpoints on the ENV tag will trigger after the
jump.  This makes intuitive sense, because the ENV tag does not
"change" until the jump is taken, but it may not be the most helpful
for debugging, since a user has to backtrack to find which instruction
caused the ENV tag change.

Change-Id: Id83494d389fb614f89b919a5c1d9245bba7a7a4e